% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_optimal_group.R
\name{agent_response}
\alias{agent_response}
\title{Get response to agents.}
\usage{
agent_response(coef_agents, coef_agents_sq, stepsize = 2, plot = FALSE)
}
\arguments{
\item{coef_agents}{The agent preference.}

\item{coef_agents_sq}{The agent count squared preference.}

\item{plot}{Whether to plot.}
}
\value{
A vector of group size preferences.
}
\description{
Get response to agents.
}
