% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_wt_lower.R
\name{cut_wt_lower}
\alias{cut_wt_lower}
\title{Scale weights, TANH transform, cut, and return numeric.}
\usage{
cut_wt_lower(x, scale = 20, min_w_val = -1.01, max_w_val = 1.01, steps = 50)
}
\arguments{
\item{x}{The values to handle.}

\item{scale}{The factor by which to multiply raw weights.}

\item{min_w_val}{The max value.}

\item{max_w_val}{The min value.}

\item{steps}{The steps between max and min values.}
}
\value{
The LOWER value of the cut bins.
}
\description{
Scale weights, TANH transform, cut, and return numeric.
}
