% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_rand_values}
\alias{get_rand_values}
\title{Test seed and random number generation.}
\usage{
get_rand_values(nValues, increment)
}
\arguments{
\item{nValues}{How many steps, each of \code{increment} magnitude, to take.}

\item{increment}{The increment in the X coordinate.}
}
\description{
Test seed and random number generation.
}
