% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLandscape}
\alias{getLandscape}
\title{Get samples from the Perlin noise landscape.}
\usage{
getLandscape(landsize, octaves, frequency, increment, clamp)
}
\arguments{
\item{landsize}{The landscape size.}

\item{octaves}{Number of octaves. May be thought of as small scale 
variability. Must be an integer value between 1 and 8. Higher values
result in landscapes with more small scale noise.}

\item{frequency}{Frequency of noise. May be thought of as large scale
variability. May be any double value between 1.0 and 16.0. Higher values
mean more patchy landscapes.}

\item{increment}{The distance between samples. Landscape variabiltiy is scale dependent,
so sampling distane must be the same when comparing between landscapes.}

\item{clamp}{The landscape value below which to return 0, useful when landscape has negative values.}
}
\description{
Returns a matrix of landscape values.
}
