% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_handle_rcpp_out.R
\name{handle_rcpp_out}
\alias{handle_rcpp_out}
\title{Process Rcpp output.}
\usage{
handle_rcpp_out(obj)
}
\arguments{
\item{obj}{The Rcpp output object, usually a list.}
}
\value{
A dataframe of individuals and weight values.
}
\description{
Process Rcpp output.
}
