% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_optimal_group.R
\name{inflection_point}
\alias{inflection_point}
\title{Get optimal group sizes.}
\usage{
inflection_point(coef_agents, coef_agents_sq)
}
\arguments{
\item{coef_agents}{The agent preference.}

\item{coef_agents_sq}{The agent count squared preference.}
}
\value{
A vector of optimal group sizes
}
\description{
Get optimal group sizes.
}
