% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_jobs_noisewalker.R
\name{make_jobs}
\alias{make_jobs}
\title{Run noisewalker on Peregrine.}
\usage{
make_jobs(
  octaves = 1L,
  f,
  ft,
  s_range,
  genmax = 10000,
  tmax = 100,
  n_rep = 3,
  ssh_con = "some server",
  password = "your password"
)
}
\arguments{
\item{octaves}{The integer number of octaves of the landscape, representing
small scale variability. This function is \emph{NOT WORKING}.
Set to a default value of 1. It is unclear what the \code{fastnoise}
default actually is.}

\item{f}{Frequency of the ancestral landscape.}

\item{ft}{Frequency of the transplant landscape.}

\item{s_range}{Individual sensory range.}

\item{genmax}{Generations on each landscape.}

\item{tmax}{Timesteps per generation.}

\item{n_rep}{How many replicates to run.}

\item{ssh_con}{The remote machine to connect to.
This should be of the form 'p123456@peregrine.hpc.rug.nl'}

\item{password}{The SSH connection password.}
}
\value{
Runs the noisewalker simulation on Peregrine or another cluster.
}
\description{
Run noisewalker on Peregrine.
}
