\name{noisewalker-package}
\alias{noisewalker-package}
\alias{noisewalker}
\docType{package}
\title{\packageTitle{noisewalker}}
\description{\packageDescription{noisewalker}}
\section{Package Content}{\packageIndices{noisewalker}}
\author{\packageAuthor{noisewalker}}
\section{Maintainer}{\packageMaintainer{noisewalker}}
\keyword{package}
