% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runNoisewalker}
\alias{runNoisewalker}
\title{Runs the noisewalker simulation.}
\usage{
runNoisewalker(genmax, timesteps, nOctaves, frequency, newSrange, rep)
}
\arguments{
\item{genmax}{The maximum number of generations per simulation.}

\item{timesteps}{The number of timesteps per generation.}

\item{nOctaves}{Number of octaves. May be thought of as small scale 
variability. Must be an integer value between 1 and 8. Higher values
result in landscapes with more small scale noise.}

\item{frequency}{Frequency of noise. May be thought of as large scale
variability. May be any double value between 1.0 and 16.0. Higher values}

\item{newSrange}{The sensory range of the population.}

\item{rep}{The replicate number. Designed to be read from a data.frame.
result in noise that is closer to white noise.}
}
\value{
Nothing. Writes an image to file.
}
\description{
Run the noisewalker simulation using parameters passed as
arguments to the corresponding R function.
}
