% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_noisewalker}
\alias{run_noisewalker}
\title{Runs the noisewalker simulation.}
\usage{
run_noisewalker(
  popsize,
  genmax,
  timesteps,
  perception,
  directions,
  costMove,
  freqRes,
  landsize,
  clamp,
  random_traits,
  scenario,
  pTransmit,
  costInfection,
  costPredAvoid,
  recordPos
)
}
\arguments{
\item{popsize}{The population size.}

\item{genmax}{The maximum number of generations per simulation.}

\item{timesteps}{The number of timesteps per generation.}

\item{perception}{The perception range.}

\item{directions}{The number of points at which agents sense resources,
at a fixed distance of \code{perception} units away from them.}

\item{costMove}{The cost per move; distance moved is assumed constant.}

\item{freqRes}{Frequency of noise. May be thought of as large scale
variability. May be any double value between 1.0 and 16.0. Higher values
mean more patchy landscapes.}

\item{landsize}{The size of the landscape after which movement is wrapped.
Actually, the landscape is infinite, and does not have a repeating pattern.}

\item{clamp}{The threshold value of the landscape below which, the agents
sense and receive zero resources. Needed because noise has values -1 to +1.}

\item{random_traits}{Should traits be initialised -1 to +1 or at 0.}

\item{scenario}{The pathogen co-evolution scenario. Defaults to 0, "none",
and no pathogen is introduced. If 1, "ancestral", the pathogen is introduced
from generation zero, the start of the simulation. If 2, "spillover", the
pathogen is introduced after 2/3 of the simulation.}

\item{pTransmit}{The probability a disease transmits.}

\item{costInfection}{The per-timestep cost of being infected.}

\item{costPredAvoid}{The cost of predation avoidance.}

\item{recordPos}{Whether to record the final position of individuals.}
}
\value{
A dataframe of evolved pop strategy count.
}
\description{
Run the noisewalker simulation using parameters passed as
arguments to the corresponding R function.
}
