% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_noisewalker}
\alias{run_noisewalker}
\title{Runs the noisewalker simulation.}
\usage{
run_noisewalker(popsize, genmax, timesteps, nOctaves, frequency, landsize)
}
\arguments{
\item{popsize}{The population size.}

\item{genmax}{The maximum number of generations per simulation.}

\item{timesteps}{The number of timesteps per generation.}

\item{nOctaves}{Number of octaves. May be thought of as small scale 
variability. Must be an integer value between 1 and 8. Higher values
result in landscapes with more small scale noise.}

\item{frequency}{Frequency of noise. May be thought of as large scale
variability. May be any double value between 1.0 and 16.0. Higher values
mean more patchy landscapes.}

\item{landsize}{The landscape size.}
}
\value{
A dataframe of evolved pop strategy count.
}
\description{
Run the noisewalker simulation using parameters passed as
arguments to the corresponding R function.
}
