% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_noisewalker}
\alias{run_noisewalker}
\title{Runs the noisewalker simulation.}
\usage{
run_noisewalker(
  popsize,
  genmax,
  timesteps,
  perception,
  directions,
  costMove,
  costSensing,
  costCompete,
  nOctaves,
  frequency,
  landsize,
  clamp
)
}
\arguments{
\item{popsize}{The population size.}

\item{genmax}{The maximum number of generations per simulation.}

\item{timesteps}{The number of timesteps per generation.}

\item{perception}{The perception range.}

\item{directions}{The number of points at which agents sense resources,
at a fixed distance of \code{perception} units away from them.}

\item{costMove}{The cost per move; distance moved is assumed constant.}

\item{costSensing}{The cost per sensing event.}

\item{costCompete}{The cost per neighbour within the perception range.}

\item{nOctaves}{Number of octaves. May be thought of as small scale 
variability. Must be an integer value between 1 and 8. Higher values
result in landscapes with more small scale noise.}

\item{frequency}{Frequency of noise. May be thought of as large scale
variability. May be any double value between 1.0 and 16.0. Higher values
mean more patchy landscapes.}

\item{landsize}{The size of the landscape after which movement is wrapped.
Actually, the landscape is infinite, and does not have a repeating pattern.}

\item{clamp}{The threshold value of the landscape below which, the agents
sense and receive zero resources. Needed because noise has values -1 to +1.}
}
\value{
A dataframe of evolved pop strategy count.
}
\description{
Run the noisewalker simulation using parameters passed as
arguments to the corresponding R function.
}
